Moved versioning tests to its own dedicated file.
